Save diff of non-__all__ attributes across import changes
This commits the diff of the output of the modattrs.py script
between when the script was introduced in 1e5a944, and now (with
the import refactoring finished and no wildcard imports remaining
outside the test suite, and only there in one import statement for
test helpers).

Neither this diff nor modattrs.py itself will be retained. Both
will be removed in the next commit. This is committed here to
facilitate inspection and debugging (especially if turns out there
are thus-far undetected regressions).