Add imports so pyright recognizes refs and index
pyright still reports git.util as private, as it should.

(mypy does not, or does not by default, report private member
access. GitPython does not generally use pyright as part of
development at this time, but I am checking some code with it
during the process of writing the new tests.)