Remove the non-ARM macOS CI jobs
This keeps only the macos-14 jobs, which run on Apple Silicon M1,
and removes the macos-13 jobs, which ran on x86-64.

Other operating systems jobs continue to run on x86-64 machines
(and none on ARM, yet). Only the macOS jobs are removed.

This change leaves Python 3.7 without any macOS test job. That is
probably okay, since it has been end-of-life for some time, and
it remains tested on Ubuntu and Windows.