Use the macos-latest label rather than macos-14
Currently they are the same. The macos-latest label will move to
later versions automatically in the future, like the ubuntu-latest
and windows-latest labels that we are already using.

In this repo, the macos-14 label had been used originally because
it was added before the migration of macos-latest to be macos-14
was completed. See https://github.com/github/roadmap/issues/926. It
was kept for clarity of constrast with the macos-13 jobs that were
also in use, some for the same Python versions.

Now that the macos-13 jobs have been removed in c2bbaf4, the
macos-latest label can be used here without confusion.