Special-case Python 3.7 for OSes we can install it on
This is analogous to the 3.7-related CI change in gitdb that was
part of https://github.com/gitpython-developers/gitdb/pull/114, as
to part of https://github.com/gitpython-developers/smmap/pull/58.

Since some tests are not yet passing on 3.13, this does not add
3.13 to CI, nor to the documentation of supported versions in
`setup.py`. Note that the list there is not enforced; GitPython can
already be installed on Python 3.13 and probably *mostly* works.

(See https://github.com/gitpython-developers/GitPython/pull/1955
for details on other changes that should be made to fully support
running GitPython on Python 3.13.)